Moved follow service to relevant directory
