+++ src/app/trade-creator-page/trade-creator/trade-creator.component.ts
@@ -11,7 +11,7 @@ import { BackendApiService } from "../../backend-api.service";
 import { ActivatedRoute, Router } from "@angular/router";
 import { CreatorCoinTrade } from "../../../lib/trade-creator-page/creator-coin-trade";
 import { AppRoutingModule } from "../../app-routing.module";
-import { FollowService } from "../../follow-button/follow.service";
+import { FollowService } from "../../../lib/services/follow/follow.service";
 
 @Component({
   selector: "trade-creator",
